moving ebay-specific code to ebay-specific scripts
